Move type to inside api directory
